separate measurement functions to separate files
